+++ packages/contentful--create-contentful-app/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 4.1.1 (2026-07-24)
+
+### 🧱 Updated Dependencies
+
+- Updated @contentful/app-scripts to 4.1.1
+
+## 4.1.0 (2026-07-23)
+
+### 🧱 Updated Dependencies
+
+- Updated @contentful/app-scripts to 4.1.0
+
 # 4.0.0 (2026-07-02)
 
 ### 🚀 Features